backend/requirements.txt
fastapi
uvicorn
pdfplumber
python-docx
nltk
spacy
scikit-learn
python-multipart
pydantic